lalala

+++ src/main/java/com/vanda/tlzbfz/controller/TRwlrController.java
@@ -5,9 +5,8 @@ import com.google.gson.Gson;
 import com.vanda.tlzbfz.common.util.GsonUtil;
 import com.vanda.tlzbfz.common.util.RedisUtil;
 import com.vanda.tlzbfz.common.util.ResultMsg;
-import com.vanda.tlzbfz.entity.RwlrExtendBean;
-import com.vanda.tlzbfz.entity.TDbrwBean;
-import com.vanda.tlzbfz.entity.TRwlrBean;
+import com.vanda.tlzbfz.entity.*;
+import com.vanda.tlzbfz.service.TBjcjsService;
 import com.vanda.tlzbfz.service.TRwlrService;
 import com.vanda.tlzbfz.service.VDbrwService;
 import io.swagger.annotations.Api;
@@ -36,6 +35,8 @@ public class TRwlrController {
     private GsonUtil gsonUtil;
     @Autowired
     private RedisUtil redisUtil;
+    @Autowired
+    private TBjcjsService bjcjsService;
 
     private static final Logger log = LoggerFactory.getLogger(TRwlrController.class);
 
@@ -45,10 +46,13 @@ public class TRwlrController {
     @PostMapping("/rwlr")
     public ResultMsg insertRwlr(@RequestBody String json,@RequestHeader("accept_token") String accept_token) throws Exception {
         try{
+            SystemLoginUser user = (SystemLoginUser) redisUtil.get(accept_token);
+            String[] unit = user.getUnitCode();
+            TBjcjs bjcjs = bjcjsService.selectBjcjs(unit[0]);
             Gson gson = gsonUtil.createGson();
             RwlrExtendBean record = gson.fromJson(json,RwlrExtendBean.class);
             //获取前端传过来的代办任务
-            List<TDbrwBean> list = record.getTdbrwBeans();
+            List<TDbrw> list = record.getTdbrwBeans();
 
             //获取前端传过来的录入bean类
             TRwlrBean rwlrBean = new TRwlrBean();
@@ -57,19 +61,24 @@ public class TRwlrController {
             ResultMsg resultMsg=new ResultMsg();
             String id = String.valueOf(System.currentTimeMillis());
             rwlrBean.setRwbh(id.substring(6,id.length()));
+            rwlrBean.setFbdw(bjcjs.getJsdm());
             int result = tRwlrService.insertSelective(rwlrBean);
 
-            for (int i=0;i<list.size();i++){
-                TDbrwBean dbrwBean = new TDbrwBean();
-                dbrwBean.setId(String.valueOf(System.currentTimeMillis()));
-                dbrwBean.setRwbh(rwlrBean.getRwbh());
-                dbrwBean.setBjcjs(list.get(i).getBjcjs());
-                dbrwBean.setZt("0");
-                dbrwBean.setWccs((long) 0);
-                dbrwBean.setSycs(rwlrBean.getYqcs());
-                dbrwBean.setGw(list.get(i).getGw());
-                vDbrwService.inserDbrwSelective(dbrwBean);
+            if(list.size()>0){
+                for (int i=0;i<list.size();i++){
+                    TDbrw dbrwBean = new TDbrw();
+                    dbrwBean.setId(String.valueOf(System.currentTimeMillis()));
+                    dbrwBean.setXm(list.get(i).getXm());
+                    dbrwBean.setRwbh(rwlrBean.getRwbh());
+                    dbrwBean.setBjcjs(list.get(i).getBjcjs());
+                    dbrwBean.setZt("0");
+                    dbrwBean.setWccs((long) 0);
+                    dbrwBean.setSycs(rwlrBean.getYqcs());
+                    dbrwBean.setGw(list.get(i).getGw());
+                    vDbrwService.inserDbrwSelective(dbrwBean);
+                }
             }
+
             if(result<=0){
                 resultMsg.setCode("400");
                 resultMsg.setMessage("录入任务失败");
@@ -94,14 +103,29 @@ public class TRwlrController {
     @Transactional(rollbackFor = Exception.class)
     @ApiOperation(value = "根据任务编号删除一条记录接口", httpMethod = "DELETE")
     @DeleteMapping("/rwlr")
-    public ResultMsg selectByRwbh(@RequestParam("rwbh") String rwbh){
+    public ResultMsg selectByRwbh(@RequestParam("rwbh") String rwbh,@RequestHeader("accept_token") String accept_token){
         try {
-            int rwlrBean = tRwlrService.deleteByRwbh(rwbh);
             ResultMsg rMsg=new ResultMsg();
+            SystemLoginUser user = (SystemLoginUser) redisUtil.get(accept_token);
+            String[] unit = user.getUnitCode();
+            TBjcjs bjcjs = bjcjsService.selectBjcjs(unit[0]);
+            List<TRwlrBean> list = tRwlrService.getRwlrByCondition(bjcjs.getJsdm(),rwbh);
+            if(list.size()==0){
+                rMsg.setCode("404");
+                rMsg.setMessage("删除记录不存在!");
+                return rMsg;
+            }
+
+            int rwlrBean = tRwlrService.deleteByRwbh(rwbh);
+
             if(rwlrBean<=0){
                 rMsg.setCode("400");
                 rMsg.setMessage("删除记录为空");
             }else {
+                List<TDbrw> dbrwBeanList = vDbrwService.selectByRwbh(rwbh);
+                for (int i=0;i<dbrwBeanList.size();i++){
+                    vDbrwService.deleteById(dbrwBeanList.get(i).getId());
+                }
                 rMsg.setCode("200");
                 rMsg.setMessage("删除成功");
             }
@@ -122,19 +146,25 @@ public class TRwlrController {
     @PutMapping("/rwlr")
     public  ResultMsg updateRwlr(@RequestBody String json,@RequestHeader("accept_token") String accept_token){
         try {
+            SystemLoginUser user = (SystemLoginUser) redisUtil.get(accept_token);
+            String[] unit = user.getUnitCode();
+            TBjcjs bjcjs = bjcjsService.selectBjcjs(unit[0]);
             Gson gson = gsonUtil.createGson();
-            TRwlrBean record = gson.fromJson(json,TRwlrBean.class);
             ResultMsg rMsg=new ResultMsg();
+            TRwlrBean record = gson.fromJson(json,TRwlrBean.class);
+            if(bjcjs.getJsdm().equals(record.getFbdw())==false){
+                rMsg.setCode("404");
+                rMsg.setMessage("更新失败！只允许发布单位用户更新数据!");
+                return rMsg;
+            }
             int rel = tRwlrService.updateBySelective(record);
             if(rel<=0){
                 rMsg.setCode("201");
-                rMsg.setMessage("任务编辑失败");
+                rMsg.setMessage("任务录入编辑失败");
             }else {
                 rMsg.setCode("200");
-                rMsg.setMessage("任务编辑成功");
-                rMsg.setData(record);
+                rMsg.setMessage("任务录入编辑成功");
             }
-
             return rMsg;
         }catch(Exception e){
             e.printStackTrace();
@@ -151,8 +181,12 @@ public class TRwlrController {
     //录入列表  条件查询  都用该接口
     @ApiOperation(value = "录入查询接口，传参为空查询所有记录，传入任务编号查询一条记录", httpMethod = "GET")
     @GetMapping("/rwlrs")
-    public ResultMsg queryRwlrList(String rwbh){
-        List<TRwlrBean> list = tRwlrService.getRwlrByCondition(rwbh);
+    public ResultMsg queryRwlrList(String rwbh,@RequestHeader("accept_token") String accept_token){
+        SystemLoginUser user = (SystemLoginUser) redisUtil.get(accept_token);
+        String[] unit = user.getUnitCode();
+        TBjcjs bjcjs = bjcjsService.selectBjcjs(unit[0]);
+        String jsmc = bjcjs.getJsdm();
+        List<TRwlrBean> list = tRwlrService.getRwlrByCondition(jsmc,rwbh);
         if(list==null){
             return  new ResultMsg("400","录入任务查询为空",null);
         }
